(maint) Expect unused filebucket to be absent
In Ruby 1.8.7 `Find.find` will ignore all given paths that are not
present, while Ruby 1.9.1+ `Find.find` will raise an exception if any of
the given paths are not present.

Before 6ed43ec a test could create an empty filebucket hierarchy, which
was then inspected to ensure it had no files. 6ed43ec fixed the
underlying cause which stopped creating the filebucket, and trying to
call Find.find on the absent directory on Ruby 1.9.1+ will fail.

This commit replaces the `Find.find` invocation with an assertion that
the filebucket directory is absent.