(#6112) Confine puppet cert generate case 3 test on windows
Windows machines are not supported as puppet master, and so should not
have puppet cert commands run on them, since they run in :master mode
and attempt to generate CA ssl infrastucture.  In particular the
ssl/ca/serial file cannot currently be handled on windows machines due
to how Puppet::Settings#readwritelock attempts to overwrite the lock
file.

This change confines this portion of the acceptance test not to run on
Windows.