Refactor to remove 'LogPath' module
In preparation for #21805, I stumbled across some coded that
seemed unnecessarily complex.  Since my changes for #21805
will end up interacting with some of this code, I decided
it was worth a quick refactor.  This commit removes the
`Puppet::Util::LogPaths` module and simplifies all of the
code that was interacting with it.