(PE-945) Add containment path to report resource_status
This commit adds a new attribute the `Puppet::Resource::Status`
class, called `containment_path`.  The value is an array of
strings that indicates the types and classes that the resource
is contained within.  (The format is the same as what we show
in the log messages during an agent run.)

This data will be used by PuppetDB to allow queries by class,
which will return resources that are contained within the
specified class.