(maint) Use systest's dynamic provisioning instead of backline
Use vcloud hypervisor with redhat template, and track puppet-acceptance
master HEAD.  Also fixed test/external_ca/fixtures file locations.

Changed the certificate validation setup step because the dynamically
provisioned hosts have an fqdn, but the reference to the hosts in
systest resolves to just the hostname which caused the master's
certificate to be only its hostname and not the fqdn.