(#21252) Define behavior around undef and nil arguments
The behavior for an argument that was :undef (from an undefined
variable) and nil (from a function returning a nil in a data structure)
was different. This fixes that inconsistency by making them both the
same. It also clarifies that passing undef to a lambda does not result
in the default value for that parameter.

In addition some of the variables have been renamed to clarify their
purpose and extra code for a 1.8.7 difference that we were unable to
reproduce has been removed.