(#21961) Support ssl connections without peer verification
This patch extends the functionality of the
  Puppet::Network::HTTP::Connection
class to support ssl connections without peer verification.
The signature of the initialize method of the class is changed
to take the "options" hash argument instead of a list of individual
arguments for easier use as well as easier addition of possible
additional parameters in the future.
The patch also changes the
  Puppet::Network::HttpPool.http_instance
method, but in a way which ensures the method signature remains
backwards compatible by only adding one parameter with a default
value.