(maint) Make DiagnosticFormatter handle diagnostics without source pos.
The problem was that DiagnosticFormatter has only been tested with the
parser (which always includes a source position even if source pos has
not information). When dealing with other forms of diagnostics there may
not be a source pos.

This fixes the formatter to not query the detailed source_pos unless it
is available.