(21985) pmt should include exsiting metadata.json when building
pmt currently will only include specific fields from the ModuleFile when
building a module. This means that as we add features to forge that rely on new
fields those won't be available to users with older versions of PMT. By allowing
metadata.json to be written to directly we ensure that module authors can take
advantage of all forge features without updating puppet. All existing workflows
continue to work for those who do not want to leverage these features.