(#21953) Add test to verify module permissions
This patch adds a call to the helper method
assert_module_installed_on_disk
in the basic module install test.

The module helper module_utils.rb is added to support it. The
criteria for validating that the module has been properly applied
to disk are as follows:

    * The containing directory for the module has a mode of 755
    * The Modulefile for hte module has a mode of 644
    * The owner of the containing directory for the module is
      the same as the `moduledir` that it is installed into.
    * The group of the containing directory for the module is
      the same as the `moduledir` that it is installed into.

The assert_module_installed_on_disk validates the above criteria.