(#21971) Add test for accessing based on master being at root
The master process, when it daemonized, switches the current working
directory to /. This means that all paths are easily knowable by a user
and no directory traversal is needed to get to any file on the file
system. The type loading code does not protect from this by making sure
that any types that are requested only appear in the known-good
directories (the manifest dirs in the modulepath) and so can be caused
to load code from anywhere on the system.