(#16856) Make a hiera.yaml in confdir be ignored if it not version 2.
The confdir-hiera scheme no has logic to skip a hiera-1 hiera.yaml
if it cannot be loaded.

This commit also fixes a problems with re-registration of injector
boot extensions. They were made in a static way causing the next
run to again register a duplicate set of bindings which caused a
conflicting bindings error. The boot bindings (for the request) should
not be system wide, and should therefore be associated with the
environment and overwritten on each request. The new mechanism changes
how the bindings from the binder_conf.yaml gets contributed to the boot
injector.