(#16856) Fix problem with YAML spec test on Ruby 1.8.7
In Ruby 1.8.7, the YAML implementation does odd things on syntax
errors in yaml source input (it is not reported and ends up being
interpted the wrong way). The test that checks for YAML syntax error is
now modified to handle RUBY 1.8.7 braindead YAML by checking for a
different issue code from the validator.