Improving testing around PMT module install permissions.
Previously, the tests simply asserted that certain files had correct permissions
and belonged to the appropriate users and groups.  This worked well enough in common
cases, but could also yield a number of false positives.  This change works to remove
a great deal of the ambiguity by specifically assigning an unusual group to the modules
directory, and by asserting the permissions for every file in the installed module.