(#21839) Stringify the :_timestamp fact name
Change all instances of the `:_timestamp` fact to `_timestamp`. Symbols can be
converted to strings during a round-trip through a simple terminus such as
JSON. If the Fact is then updated using the symbol, there will be two names for
the same data.

After 5ac681f, in the 3.0 rc series, all varible names are explicitly
stringified when added to compiler scopes. This means that facts named
`:_timestamp` and `_timestamp` will end up mapping to the same scope variable
which violates static single assignment.