(Maint) Improve check for the timestamp
Since the test cannot control how much time elapses between getting a
time and the timestamp being created, the test was brittle as written
(that the timestamp should be within 1 second of the time at the start
of the test). The right way to check this is to assert that the
timestamp is from some time between the start of the operation and the
end of the operation.