(Maint) Collapse classes for relationship graphs
Since the introduction of the Puppet::RelationshipGraph the
Puppet::Transaction::RelationshipGraphWrapper seemed a bit out of place.
This changes the interface for the #traverse method so that all of the
transaction specific and interacting portions are injected in, which
decouples the RelationshipGraphWrapper from the Transaction. Once that
was done, then it was clear that the two graph classes could be brought
together.

This provides a single class that represents the final relationship
graph and the semantics of traversal that that graph has.