(#19886) Change syntax check feature to use extensions/injector
This changes the heredoc feature to perform syntax checking to
lookup available syntax checkers via Puppet Bindings instead of
via special purpose parser functions.

This is good because the functions were not intended to be used from user
logic as one function per "syntax" was required. If a function is wanted
it should be a generic function that looks up the syntax.