(#22128) Fixes rpm provider's nevra regex to work in yum prefetch
Work done for #19875 changed the rpm provider's regex for parsing
package information.  The yum package provider's prefetch method relied
on the earlier version of this regex and was broken by this change,
resulting in a failure to set latest_info when using yum and a series of
warning messages in the logs.

This change adjusts the rpm NEVRA_REGEX such that the added description
info is an optional capture.