(maint) Load puppetx for yaml_backend_spec
Depending on order of specs, dependencies might or might not be loaded
for Puppet::Pops::Binder::Hiera2::YamlBackend to resolve.

Ensures that YamlBackend's prerequisites are loaded in the spec.
PuppetX will only load the PuppetX::Puppet::Hiera2Backend that
YamlBackend extends if binder is true.

This is not a particularly lovely way of doing this in a spec, but the
YamlBackend implementation is intended to be lazily loaded by the
binding system.  There may be a better way of ensuring YamlBackend is
available for testing.