(#22128) Yum prefetch spec ensures use of yum provider
The prefetch spec for the Yum package provider was failing on platforms
whose default providers did not have the same capabilities (definitely
on Windows for msi), because the previous implementation was generating type
instances of Puppet::Type::Package without setting their providers.
Consequently these instances were being generated with references to the
platform's default providers, and might fail to set :ensure =>
:latest due to missing capabilities.

The spec is changed to associate instances of yum providers with the
package type instances being passed into prefetch.  Also removed the
confine for Windows.

This also clears up a typo that got through in the original patch, which
was failing on only some platforms.