(#18508) Handle duplicate generated resources correctly
When a generated resource duplicates a resource already in the catalog,
we want to avoid adding it a second time, but we still want to create
dependencies between the resource and the generating resource. For
example, we may have a recursive file resource, but explicitly model one
of the files that will be generated. We still want the explicit file
resource to depend on its generated parent directory.

This also refactors the resource generation methods, splitting them up
into smaller methods.