(Maint) Move most graph related code to Puppet::Graph
We often talk about "the graph", but until now there hasn't been a clear
graph that we were talking about. The relationship graph that is used
for evaluation is the best candidate to fill this role, since it
contains all of the ordering information that is needed.

This also removes Puppet::Util::Graph, which was unused.