(#22205) Extract class for generating resource priorities
Previously the code for generating priorities was scattered throughout
the places that needed to add vertices to the relationship graph. This
pulls out that logic into a class that can generate the priorities for
resources, and those contained in other resources. This is needed to be
able to make this strategy pluggable.