(Maint) Restructure requires for settings
The settings classes were not loaded in a robust manner. Some required
other files with the knowledge that the right modules would be created.
This caused a duplication of requires (which can cause a small slowdown
as ruby looks up whether the code has already been loaded). By requiring
the settings *after* the settings class has been opened, we can avoid a
lot of this extra work and overhead.