(maint) Refactor requirement on rspec helpers for pops
The helper classes for rspec testing of pops related functionality was
included using __FILE__ relative paths instead of being placed in a
directory on the riby load path.

This commit moves the helpers (factory_rspec_helper,
parser_rspec_helper, and transformer_rspec_helper) and changes the
require statements to use a location under 'puppet_spec' in the test
that makes use of these helpers.