(#22432) Sort defaults/commands/features in provider doc hooks
Because hashes are unordered in Ruby, these lists would churn around randomly
each time you generated the references, making it harder to mechanically compare
two versions of the type reference. Also, if the list is long and you're
actually looking for something in it, that would make it harder to scan. So
let's sort the arrays returned by these collect calls!