(#22288) Add methods to access all parameters of a type
Without this commit there is no simple way to access all parameters of
a type. There is also no simple way to iterate over them. There is
currently a method that will return a list of all parameters that have a
value assigned, however, they are duplicates of the actual parameters. This
means there is no clear way to perform an operation on all parameters at once
that will have lasting side effects.

This commit adds functionality needed by the solution to ticket
(#22288).This functionality this commit adds includes a method to collect all
parameters that currently have a value assigned. This commit also adds a method
to iterate of the collected parameters that currently have a value. The intention
of these methods is to allow things like post catalog compilation hooks to be possible.