(#22288) Add post_compile_hook to Puppet::Parameter
Without this commit there is no message sent to parameters (and
subclasses thereof) to indicate the catalog has completed compilation.
This is an issue for any parameter that is not based on
`Puppet::Type::RelationshipMetaparam` and needs to do something on
completion of catalog compilation. An example for such application would
be verifying some other resource made it into the catalog so that a
value may be used fro it for the current type.

This commit add an enhancement to the `Puppet::Type#finish` method. This
method will now list through all parameters, including
`Puppet::Type::RelationshipMetaparam` parameters, and call the
`post_compile_hook` method if implemented. This commit also adds an
abstract method comment to `Puppet::Parameter`.