(#22288) Add tests for new methods in `Puppet::Type`
Without this commit there is no way to test the functionality
added to access all parameters in `Puppet::Type`. The functionality
being tested consists of two new methods to access all parameters that
have values set for an instance of `Puppet::Type`.

This commit adds three tests to the 'spec/unit/type_spec.rb' to test the
new behavior.