(#22288) Add test coverage for `post_compile_hook`
Without this commit there is no way to test the functionality surrounding
the `post_compile_hook' for parameters. The functionality mentioned allows a
`post_compile_hook` method to be implemented by any parameter. This hook is
called on the parameters from the `Puppet::Type#finish` method.

This commit adds a new test to 'spec/unit/type_spec.rb' in order to validate
the new behavior that has been added.