(#22319) Correct Puppet::Setting's use of WatchedFile
WatchedFile was introduced after 3.2.0 as a new method of checking for
changes to file.  Puppet::Settings was calling a non-existent method on
the new WatchedFile, and would cause a daemonized Puppet to abort if one
of its watched configuration files was changed.

This is a minor change to call #to_str instead.